RESTRUCTURE script utils a bit
