+++ scripts/utils/cli-utils.js
@@ -1,6 +1,4 @@
-#!/usr/bin/env node
-
-/* eslint-disable global-require, no-octal-escape */
+/* eslint-disable import/no-extraneous-dependencies, global-require, no-octal-escape */
 const childProcess = require('child_process');
 
 const logger = console;
