+++ scripts/utils/watch-babel.js
